requirements.txt
asgiref==3.8.1
Django==5.0.4
django-resized==1.0.2
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
Unidecode==1.3.8
uritemplate==4.1.1
